+++ CHANGES.md
@@ -2,6 +2,7 @@ CHANGELOG
 =======
 
 ##Version 0.1.8
+**May 20, 2014**
 
 - Better handling of terminal colors 
 - Send keynames to LS when updated, to record the modified stamp